requirements.txt
python-telegram-bot[webhooks]==22.5
flask==3.0.3
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0